+++ hasklepias-core/src/Hasklepias/MakeFilterApp.hs
@@ -62,7 +62,7 @@ newtype FilterAppOpts = FilterAppOpts
 makeAppArgs :: String -> ParserInfo FilterAppOpts
 makeAppArgs name = Options.Applicative.info
   (FilterAppOpts <$> (fileInput <|> s3Input <|> stdInput) <**> helper)
-  (fullDesc <> header ("Filter events for " <> name))
+  (fullDesc <> progDesc desc <> header ("Filter events for " <> name))
 
 -- A type to hold a subject ID
 newtype HoldID = MkID SubjectID deriving (Eq, Show)
@@ -141,7 +141,31 @@ prefilterC p x =
 -- | Type containing the filter app
 newtype FilterApp m = MkFilterApp { runPreApp :: Maybe Location -> m C.ByteString }
 
+desc = 
+  "The application takes event data formatted as ndjson (http://ndjson.org/) \
+  \(i.e. one event per line). The application returns the event data filtered to \
+  \all those subjects who have at least one event satisfying the given predicate. \
+  \Each subject's data must be grouped in contiguous chunks of lines; otherwise, \
+  \the application may not behave as expected and will not warn or raise an error. \
+  \Lines that fail to parse as an `Event` do not satisfy the predicate, but are not \
+  \dropped from the output. In other words, all of a subject's data is returned in \
+  \the same order as the input, provided that at least one line successfully parses \
+  \into an Event and satisfies the predicate." 
+
 {- | 
+Create a application that filters event data with two arguments: 
+  * a string for the name of the application (e.g. the project ID)
+  * a predicate function of type @Event a -> Bool@. 
+
+The application takes event data formatted as [`ndjson`](http://ndjson.org/)
+(i.e. one event per line). The application returns the event data filtered to
+all those subjects who have at least one event satisfying the given predicate.
+Each subject's data must be grouped in contiguous chunks of lines; otherwise, 
+the application may not behave as expected and will not warn or raise an error.
+Lines that fail to parse as an `Event` do not satisfy the predicate, but are not
+dropped from the output. In other words, all of a subject's data is returned in
+the same order as the input, provided that at least one line successfully parses
+into an `Event` and satisfies the predicate. 
 -}
 makeFilterApp
   :: (Show a, FromJSON a, IntervalSizeable a b)