+++ hasklepias-main/src/Hasklepias/ExampleFilterApp.hs
@@ -5,5 +5,6 @@ module Hasklepias.ExampleFilterApp
   ) where
 import           Hasklepias
 
+-- | A basic test of the filter app: filter to subject that have at least one enrollment event.
 exampleFilterApp :: FilterApp IO
-exampleFilterApp = makeFilterApp "test" (getPredicate isEnrollmentEvent :: Event Day -> Bool)
+exampleFilterApp = makeFilterApp "Prefilter test app" (getPredicate isEnrollmentEvent :: Event Day -> Bool)