+++ hasklepias-main/ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for hasklepias-main
 
+## 0.22.4
+
+* Adds `Hasklepias.ExampleFilterApp` module as an example application that filters subjects have at least one event that satisfies a given predicate function. 
+* Adds a test suite for the `exampleFilterApp`.
+
 ## 0.22.3
 
 * Updates examples to run with updates to `hasklepias-core`.